==> 0_00_saudistream.ipynb <==
# Import package
import tweepy
import json

consumer_key = "ENTER>consumer_key"
consumer_secret = "ENTER>consumer_secret"
access_token = "ENTER>access_token"
access_token_secret = "ENTER>access_token_secret"


BOX = [34.6323360532, 16.3478913436, 55.6666593769, 32.161008816]

# Pass OAuth details to tweepy's OAuth handler
auth = tweepy.OAuthHandler(consumer_key, consumer_secret)
auth.set_access_token(access_token, access_token_secret)


# api = tweepy.API(auth)
# api.update_status('tweepy + oauth!')

# class use json
# creates a Tweet listener that creates a file called 'sauditweets.txt'
# collects streaming tweets as .jsons and writes them to the file 'sauditweets.txt'
# once n tweets have been streamed (if self.num_tweets < n:), the listener closes the file and stops listening

class MyStreamListener(tweepy.StreamListener):
    
    def __init__(self, api=None):
        # inherit class attributes
        super(MyStreamListener, self).__init__()
        self.num_tweets = 0
        self.file = open("data/output/sauditweets.txt", "w+")

    def on_status(self, status):
        tweet = status._json
        
        self.file.write( json.dumps(tweet) + '\n' )
        
        self.num_tweets += 1
        if self.num_tweets < 100000:
            return True
        else:
            return False
        self.file.close()

    def on_error(self, status):
        print(status)

# Initialize Stream listener
l = MyStreamListener()

# Create you Stream object with authentication
stream = tweepy.Stream(auth, l)


# Filter Twitter Streams to capture data by the keywords:
stream.filter(locations=BOX)
