==> 0_01_saudistream_gencsv.ipynb <==
import json
import pandas as pd

tweets_data = []
# Open connection to file
h=open('data/output/sauditweets.txt','r')

# Read in tweets and store in list: tweets_data
for i in h:
    try:
        tmp=json.loads(i)
        tweets_data.append(tmp)
    except:
        print('X'), #shows lines with errors
h.close()

df = pd.DataFrame(tweets_data)
df

places = df['place'].apply(pd.Series)
places

places = places.rename(columns = lambda x : 'place_' + str(x))
places

df = pd.concat([df[:], places[:]], axis=1)

users = df['user'].apply(pd.Series)
users = users.rename(columns = lambda x : 'user_' + str(x))
users

df = pd.concat([df[:], users[:]], axis=1)

entities = df['entities'].apply(pd.Series)
entities

df = pd.concat([df[:], entities[:]], axis=1)
df

df.describe()

##save to csv
df.to_csv("data/output/sauditweets.csv")
